Rename app from VectoryTones to VistaTotes
- Updated all displayed names in UI and documentation
- Renamed vectorytones folder to vistatotes
- Updated all Python imports to use new package name
- Updated git remote URL
