+++ REFACTORING.md
@@ -1,11 +1,11 @@
-# VectoryTones Refactoring
+# VistaTotes Refactoring
 
-This document describes the refactored codebase structure for VectoryTones.
+This document describes the refactored codebase structure for VistaTotes.
 
 ## New Package Structure
 
 ```
-vectorytones/
+vistatotes/
 ├── __init__.py           # Package initialization
 ├── audio/                # Audio utilities
 │   ├── __init__.py
@@ -56,7 +56,7 @@ def generate_wav(frequency: float, duration: float) -> bytes:
 
 **After:**
 ```python
-from vectorytones.audio import generate_wav, wav_bytes_to_float
+from vistatotes.audio import generate_wav, wav_bytes_to_float
 ```
 
 ### Model Loading
@@ -70,7 +70,7 @@ def initialize_app():
 
 **After:**
 ```python
-from vectorytones.models import initialize_models, get_clap_model
+from vistatotes.models import initialize_models, get_clap_model
 
 initialize_models()
 clap_model, clap_processor = get_clap_model()
@@ -88,7 +88,7 @@ def embed_audio_file(audio_path: Path) -> Optional[np.ndarray]:
 
 **After:**
 ```python
-from vectorytones.models import embed_audio_file, embed_text_query
+from vistatotes.models import embed_audio_file, embed_text_query
 
 # Embed media files
 embedding = embed_audio_file(audio_path)
@@ -110,8 +110,8 @@ def train_and_score(inclusion_value=0):
 
 **After:**
 ```python
-from vectorytones.models import train_and_score
-from vectorytones.utils import clips, good_votes, bad_votes, get_inclusion
+from vistatotes.models import train_and_score
+from vistatotes.utils import clips, good_votes, bad_votes, get_inclusion
 
 results, threshold = train_and_score(clips, good_votes, bad_votes, get_inclusion())
 ```
@@ -128,7 +128,7 @@ def update_progress(status, message="", current=0, total=0, error=None):
 
 **After:**
 ```python
-from vectorytones.utils import update_progress, get_progress
+from vistatotes.utils import update_progress, get_progress
 
 update_progress("loading", "Loading models...", 1, 3)
 progress = get_progress()
@@ -146,8 +146,8 @@ inclusion = 0
 
 **After:**
 ```python
-from vectorytones.utils import clips, good_votes, bad_votes
-from vectorytones.utils import clear_all, set_inclusion, get_inclusion
+from vistatotes.utils import clips, good_votes, bad_votes
+from vistatotes.utils import clear_all, set_inclusion, get_inclusion
 
 # Use imported state
 clips[clip_id] = {...}
@@ -173,7 +173,7 @@ DEMO_DATASETS = {
 
 **After:**
 ```python
-from vectorytones.datasets import DEMO_DATASETS
+from vistatotes.datasets import DEMO_DATASETS
 
 for dataset_name, config in DEMO_DATASETS.items():
     print(f"{dataset_name}: {config['description']}")
