+++ static/index.html
@@ -3,7 +3,7 @@
 <head>
 <meta charset="UTF-8">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>VectoryTones</title>
+<title>VistaTotes</title>
 <style>
   *, *::before, *::after { box-sizing: border-box; margin: 0; padding: 0; }
   body { font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif; background: #0f1117; color: #e0e0e0; height: 100vh; display: flex; flex-direction: column; }
@@ -209,7 +209,7 @@
       </div>
     </div>
   </div>
-  <h1>VectoryTones</h1>
+  <h1>VistaTotes</h1>
   <span>Sound clip explorer with vector embeddings</span>
 </header>
 <div class="layout">
@@ -265,7 +265,7 @@
   <div class="panel-center" id="center">
     <div class="dataset-welcome" id="dataset-welcome">
       <div>
-        <h2>Welcome to VectoryTones</h2>
+        <h2>Welcome to VistaTotes</h2>
         <p>Load or generate a dataset to get started</p>
       </div>
       <div class="dataset-options" id="dataset-options">
