--- vectorytones/audio/__init__.py
@@ -1,9 +1,0 @@
-"""Audio generation and processing utilities."""
-
-from vectorytones.audio.generator import generate_wav
-from vectorytones.audio.processor import wav_bytes_to_float
-
-__all__ = [
-    "generate_wav",
-    "wav_bytes_to_float",
-]
