+++ vistatotes/audio/__init__.py
@@ -1,0 +1,9 @@
+"""Audio generation and processing utilities."""
+
+from vistatotes.audio.generator import generate_wav
+from vistatotes.audio.processor import wav_bytes_to_float
+
+__all__ = [
+    "generate_wav",
+    "wav_bytes_to_float",
+]
