+++ vistatotes/models/__init__.py
@@ -1,21 +1,21 @@
 """Model loading, embeddings, and training utilities."""
 
-from vectorytones.models.embeddings import (
+from vistatotes.models.embeddings import (
     embed_audio_file,
     embed_image_file,
     embed_paragraph_file,
     embed_text_query,
     embed_video_file,
 )
-from vectorytones.models.loader import (
+from vistatotes.models.loader import (
     get_clap_model,
     get_clip_model,
     get_e5_model,
     get_xclip_model,
     initialize_models,
 )
-from vectorytones.models.progress import analyze_labeling_progress
-from vectorytones.models.training import (
+from vistatotes.models.progress import analyze_labeling_progress
+from vistatotes.models.training import (
     calculate_cross_calibration_threshold,
     calculate_gmm_threshold,
     find_optimal_threshold,
