+++ vistatotes/models/training.py
